Signin restructure.
signin_pref_names.h and other header files used in unit tests are now in
components/signin/public/base/.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/3517170ea99d55a43b030a55409ddc300846d0f9

commit 3517170ea99d55a43b030a55409ddc300846d0f9
Author: Colin Blundell <blundell@chromium.org>
Date:   Thu Jul 11 08:16:34 2019 +0000

    Signin: Restructure //components/signin/core/browser:shared

    This CL moves the files that are currently in the
    //components/signin/core/browser:shared GN target into the structure
    described in this document:

    https://docs.google.com/document/d/1p_kahITGFgv7PehQ61-thRVR3Hrdaf3tLryMytTRPYs/edit#

    Namely, files intended for usage by the entire signin component and its
    consumers are moved into a new //signin/public/base/ directory. Files
    that are conceptually part of the IdentityManager public interface are
    moved into //signin/public/identity_manager/.

    This change was done in two parts:

    1. A script that did the code moves and corresponding renames,
       include path changes, etc.
    2. Manual followup that changed GN files and DEPS to reflect the new
       structure.

    For ease of reviewing, I uploaded 1 as the original PS and 2 as PS2.

    TBR=jochen@chromium.org

    Bug: 952788
