+++ components/brave_sync/brave_profile_sync_service_impl.cc
@@ -23,7 +23,7 @@
 #include "brave/components/brave_sync/values_conv.h"
 #include "chrome/browser/sync/chrome_sync_client.h"
 #include "components/bookmarks/browser/bookmark_model.h"
-#include "components/signin/core/browser/account_info.h"
+#include "components/signin/public/identity_manager/account_info.h"
 #include "components/sync/engine_impl/syncer.h"
 #include "content/public/browser/browser_thread.h"
 #include "net/base/network_interfaces.h"
