+++ test/BUILD.gn
@@ -244,7 +244,7 @@ test("brave_unit_tests") {
     "//components/prefs:test_support",
     "//components/version_info",
     "//content/test:test_support",
-    "//components/signin/core/browser",
+    "//components/signin/public/base",
     "//components/sync_preferences",
     "//components/translate/core/browser:test_support",
     "//content/public/common",
